Package structure
Created 3 main packages:

- Cache: REDIS cache
- Filter: filter protocol for malware URL
- Handler: handlers managing HTTP connection requests
